% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_removal_model.R
\name{abundance_removal_model}
\alias{abundance_removal_model}
\title{Abundance Removal Model
A wrapper function run JAGS-based removal models for stream data.
Takes habitat, transect, and count data, processes it, runs the models for each species, and saves the output.}
\usage{
abundance_removal_model(count, output_dir)
}
\arguments{
\item{count}{Path to the count data file (CSV format).}

\item{output_dir}{Directory to save the outputs.}

\item{hab}{Path to the habitat data file (CSV format).}

\item{tran}{Path to the transect habitat data file (CSV format)}
}
\value{
A list containing count-habitat model summaries and plots.
}
\description{
Abundance Removal Model
A wrapper function run JAGS-based removal models for stream data.
Takes habitat, transect, and count data, processes it, runs the models for each species, and saves the output.
}
