% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_removal_model.R
\name{habitat_removal_model}
\alias{habitat_removal_model}
\title{Habitat Removal Model
A wrapper function run JAGS-based removal models for stream data.
Takes habitat, transect, and count data, processes it, runs the models for each species, and saves the output.}
\usage{
habitat_removal_model(hab, tran, count, output_dir)
}
\arguments{
\item{hab}{Path to the habitat data file (CSV format).}

\item{tran}{Path to the transect habitat data file (CSV format)}

\item{count}{Path to the count data file (CSV format).}

\item{output_dir}{Directory to save the outputs.}
}
\value{
A list containing count-habitat model summaries and plots.
}
\description{
Habitat Removal Model
A wrapper function run JAGS-based removal models for stream data.
Takes habitat, transect, and count data, processes it, runs the models for each species, and saves the output.
}
